fix(docker)!: Proper image tagging 🏷️
Now the images are properly tagged. For example, the image php-dev:8 has the same digest as php-dev:8.2. As a bonus, the build workflow has been split, so only the appropriate image will be rebuilt.

BREAKING CHANGE: from now on, the "latest" tag refers to the latest image of this repository, not the latest image of the corresponding service.